requirements.txt
asn1crypto==0.24.0
attrs==18.2.0
Automat==0.6.0
bcrypt==3.1.7
blinker==1.4
certifi==2018.8.24
cffi==1.14.0
chardet==3.0.4
Click==7.0
colorama==0.3.7
configobj==5.0.6
constantly==15.1.0
cryptography==2.6.1
dbus-python==1.2.12
distro==1.5.0
entrypoints==0.3
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
gunicorn==20.0.4
httplib2==0.11.3
hyperlink==17.3.1
idna==2.6
incremental==16.10.1
itsdangerous==1.1.0
Jinja2==2.11.2
keyring==18.0.1
keyrings.alt==3.1.1
launchpadlib==1.10.7
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==1.1.1
netifaces==0.10.4
numpy==1.18.2
oauthlib==2.1.0
pandas==1.0.3
pbr==5.4.5
Pillow==7.1.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.20
pycrypto==2.6.1
PyHamcrest==2.0.2
python-dateutil==2.8.1
pytz==2019.3
pyxdg==0.25
PyYAML==5.1.2
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.12.0
SQLAlchemy==1.3.16
systemd-python==234
testresources==2.0.1
Twisted==18.9.0
urllib3==1.24.1
wadllib==1.3.3
Werkzeug==1.0.1
WTForms==2.2.1
zope.interface==4.3.2
